# knowledge-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/knowledge/relations/{id}": {
      "get": {
        "tags": ["knowledge-read-controller"],
        "operationId": "relations",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RelationsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/recent": {
      "get": {
        "tags": ["knowledge-read-controller"],
        "operationId": "listRecent",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NotesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/recall": {
      "get": {
        "tags": ["knowledge-read-controller"],
        "operationId": "recall",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RecallResultResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge/notes/{id}": {
      "get": {
        "tags": ["knowledge-read-controller"],
        "operationId": "getNote",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/knowledge/conflicts/{id}": {
      "get": {
        "tags": ["knowledge-read-controller"],
        "operationId": "conflicts",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RelationsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RelationResponse": {
        "type": "object",
        "properties": {
          "subjectId": {
            "type": "string"
          },
          "predicate": {
            "type": "string"
          },
          "objectId": {
            "type": "string"
          },
          "props": {
            "type": "object",
            "additionalProperties": {}
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": ["createdAt", "objectId", "predicate", "props", "subjectId"]
      },
      "RelationsResponse": {
        "type": "object",
        "properties": {
          "relations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelationResponse"
            }
          }
        },
        "required": ["relations"]
      },
      "NoteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "capturedAt": {
            "type": "string"
          },
          "sessionId": {
            "type": ["string", "null"]
          },
          "confidence": {
            "type": "number",
            "format": "double"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "vaultPath": {
            "type": "string"
          },
          "vaultCommit": {
            "type": ["string", "null"]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["body", "capturedAt", "confidence", "id", "scope", "source", "tags", "title", "type", "vaultPath"]
      },
      "NotesResponse": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoteResponse"
            }
          }
        },
        "required": ["notes"]
      },
      "RecallHitResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "snippet": {
            "type": "string"
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["id", "scope", "score", "snippet", "tags", "title", "type"]
      },
      "RecallResultResponse": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string"
          },
          "hits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecallHitResponse"
            }
          }
        },
        "required": ["hits", "mode"]
      },
...
    }
  }
}